Process strings.js
