+++ src/utilities/strings.js
@@ -1,29 +1,4 @@
-import isString from 'lodash.isstring'
 import { normalizeUrl } from './urls'
-
-export const isWordString = word => {
-  return !!(isString(word) && word.length)
-}
-
-export const isWord = word => {
-  return typeof word === 'number' || isWordString(word)
-}
-
-export const textIncludesOnlyEmoji = text => {
-  // Note: `\u2028`& `\uFE0F` are not emoji characters, but instead unicode line
-  // separators that seem to appear when the `replace` methods are called below.
-  const emojiPattern = /(\u2028|\uFE0F|\u00a9|\u00ae|[\u2000-\u3300]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff])/g
-
-  if (!isWordString(text)) return false
-
-  // Remove all emoji and see if any characters remain
-  const withEmojisRemoved = text.replace(emojiPattern, '')
-
-  // Remove spaces
-  const withSpacesRemoved = withEmojisRemoved.replace(/[\s\n]/gm, '')
-
-  return !withSpacesRemoved
-}
 
 export const newlineToHTML = string => {
   if (!string) {
@@ -53,21 +28,6 @@ export const repeat = (pattern, count) => {
   }
 
   return result + pattern
-}
-
-/**
- * Camelcases a specified string.
- *
- * @param   {string} string The string.
- * @returns {string} The camelCased string.
- */
-export const camelCase = string => {
-  return string
-    .replace(/-/g, ' ')
-    .replace(/(?:^\w|[A-Z]|\b\w|\s+)/g, (match, index) => {
-      if (+match === 0) return '' // or if (/\s+/.test(match)) for white spaces
-      return index === 0 ? match.toLowerCase() : match.toUpperCase()
-    })
 }
 
 // Taken from the React escapeTextForBrowser internal utility
